litigation_requirements.txt
# Requirements for Private AI Litigation Knowledge System

# Core Gradio (already in main requirements.txt)
gradio

# Document Processing
PyPDF2>=3.0.0
python-docx>=1.0.0
pandas>=1.0.0
openpyxl>=3.0.0  # For Excel file support

# Vector Embeddings and Search
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4  # Use faiss-gpu if you have CUDA
numpy>=1.0.0

# Optional: For better email processing
# extract-msg  # For .msg files (Windows Outlook)

# Optional: For advanced AI capabilities
# transformers>=4.20.0  # If you want to use local LLMs
# torch>=1.12.0  # If using transformers

# Optional: For better text extraction
# pdfplumber>=0.9.0  # Alternative PDF processor
# python-docx2txt>=0.8  # Alternative DOCX processor

